LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/email-set-value-expected.txt



PASS setValue(sanitizedValue) is reflected in visible text field content
